+++ internal/auth/auth.go
@@ -554,6 +554,39 @@ func (r *VerifyResults) AdditionalVerification(ctx context.Context, opt ...Optio
 	return
 }
 
+// FetchActionsForId returns the allowed actions for a given ID using the
+// current set of ACLs and all other parameters the same (user, scope, etc.)
+func (r *VerifyResults) FetchActionsForId(ctx context.Context, id string, availableActions action.Actions, opt ...Option) action.Actions {
+	switch {
+	case r.v.requestInfo.DisableAuthEntirely,
+		r.v.requestInfo.TokenFormat == AuthTokenTypeRecoveryKms:
+		// TODO: See if we can be better about what we return with the anonymous
+		// user and recovery KMS, perhaps given exclusionary options for each
+		return availableActions
+	}
+
+	opts := getOpts(opt...)
+	res := opts.withResource
+	if res == nil {
+		res = r.v.res
+	}
+	if res == nil {
+		res = new(perms.Resource)
+	}
+	res.Id = id
+
+	ret := make(action.Actions, 0, len(availableActions))
+	for _, act := range availableActions {
+		if r.v.acl.Allowed(*res, act).Allowed {
+			ret = append(ret, act)
+		}
+	}
+	if len(ret) == 0 {
+		return nil
+	}
+	return ret
+}
+
 // GetTokenFromRequest pulls the token from either the Authorization header or
 // split cookies and parses it. If it cannot be parsed successfully, the issue
 // is logged and we return blank, so logic will continue as the anonymous user.