+++ internal/errors/code_test.go
@@ -53,6 +53,26 @@ func TestCode_Both_String_Info(t *testing.T) {
 			want: EmptyFieldMask,
 		},
 		{
+			name: "KeyNotFound",
+			c:    KeyNotFound,
+			want: KeyNotFound,
+		},
+		{
+			name: "TicketAlreadyRedeemed",
+			c:    TicketAlreadyRedeemed,
+			want: TicketAlreadyRedeemed,
+		},
+		{
+			name: "TicketNotFound",
+			c:    TicketNotFound,
+			want: TicketNotFound,
+		},
+		{
+			name: "Io",
+			c:    Io,
+			want: Io,
+		},
+		{
 			name: "PasswordTooShort",
 			c:    PasswordTooShort,
 			want: PasswordTooShort,
@@ -71,6 +91,26 @@ func TestCode_Both_String_Info(t *testing.T) {
 			name: "PasswordsEqual",
 			c:    PasswordsEqual,
 			want: PasswordsEqual,
+		},
+		{
+			name: "Encrypt",
+			c:    Encrypt,
+			want: Encrypt,
+		},
+		{
+			name: "Decrypt",
+			c:    Decrypt,
+			want: Decrypt,
+		},
+		{
+			name: "Encode",
+			c:    Encode,
+			want: Encode,
+		},
+		{
+			name: "Decode",
+			c:    Decode,
+			want: Decode,
 		},
 		{
 			name: "CheckConstraint",