+++ internal/errors/info.go
@@ -36,6 +36,22 @@ var errorCodeInfo = map[Code]Info{
 		Message: "empty field mask",
 		Kind:    Parameter,
 	},
+	KeyNotFound: {
+		Message: "key/version not found",
+		Kind:    Integrity,
+	},
+	TicketAlreadyRedeemed: {
+		Message: "ticket already redeemed",
+		Kind:    Integrity,
+	},
+	TicketNotFound: {
+		Message: "ticket not found",
+		Kind:    Integrity,
+	},
+	Io: {
+		Message: "error during io operation",
+		Kind:    Integrity,
+	},
 	PasswordTooShort: {
 		Message: "too short",
 		Kind:    Password,
@@ -51,6 +67,22 @@ var errorCodeInfo = map[Code]Info{
 	PasswordsEqual: {
 		Message: "old and new password are equal",
 		Kind:    Password,
+	},
+	Encrypt: {
+		Message: "error occurred during encrypt",
+		Kind:    Encryption,
+	},
+	Decrypt: {
+		Message: "error occurred during decrypt",
+		Kind:    Encryption,
+	},
+	Encode: {
+		Message: "error occurred during encode",
+		Kind:    Encoding,
+	},
+	Decode: {
+		Message: "error occurred during decode",
+		Kind:    Encoding,
 	},
 	CheckConstraint: {
 		Message: "constraint check failed",