+++ internal/errors/kind.go
@@ -10,6 +10,8 @@ const (
 	Search
 	Password
 	Transaction
+	Encryption
+	Encoding
 )
 
 func (e Kind) String() string {
@@ -20,5 +22,7 @@ func (e Kind) String() string {
 		Search:      "search issue",
 		Password:    "password violation",
 		Transaction: "db transaction issue",
+		Encryption:  "encryption issue",
+		Encoding:    "encoding issue",
 	}[e]
 }